noise_filter: move from analysis/ to core/
It's part of the review pipeline (used only by core/engine.py), not a
standalone analysis tool. analysis/ was misleading placement.
