+++ src/mira/core/noise_filter.py
